Get nuStyleAttrib down to 8 bytes
Get android on par with Win32.
Hide all DomEl mutators behind functions, so that we can have a reliable
version number for each element.
Create a job queue and worker pool.
Implement SurfaceLost for android.
Fix AbcQueue so that it doesn't pretend that AbcCriticalSection is
re-enterable on linux.
Show number of CPU cores at startup.. surprisingly "3" on Samsung S3.