Optimize for less DOM churn vs cheap whole-DOM copy.
Make the renderer's clone of the document use the heap instead of
one big pool that is discarded every time the document is synced.

Being fancy about your heap allocs can only get you so far when
memory bandwidth is your most urgent constraint.

So instead we optimize for the case where large DOMs are not
recreated with every frame.