DirectX is working, except for text shaders.
One of the things I wanted out of DirectX was lower latency
on Windows 8, but it turns out the thing that introduced
latency on OpenGL on Windows 8 was simply that VSync
is on by default. Calling wglSwapIntervalEXT(0) is all that's
needed to get OpenGL latency on Windows 8 down to a
minimum, and on par with DirectX.
Nevertheless, DirectX support will enable this to run on
a broader range of Windows PCs.