FORMAT: Remove spaces inside parentheses
At IMQS (my workplace) we needed to do the same change on our C++
codebase, in the name of uniformity and principle of least surprise. Since
I am the only developer on xo, I choose to do the same thing here, because
I don't want to have to live with two different styles enforced by an IDE
etc.