Added 'stretching bind points'
If you bind to vcenter or baseline, and then also bind to top or bottom,
then the top or bottom act after the vcenter or baseline. What they do in
this case is they stretch the box to the top bind-point, and stretch it to
the bottom bind-point. If you only have a single bind-point then you move
the box instead of stretching it.
For the bvckup2 UI that I'm replicating, this works great to do deep
baseline alignment without having to specify your padding or height.