Rename build to meta, so that I can use 'build' for CMake
It's conventional to use 'build' as the CMake build folder, so I'm gonna
stick to that. Whether we like it or not, CMake has become the lingua
franca of C++ build systems, so we need to support it. What's pushed me
over the brink is the desire to have "Edit And Continue" support for GUI
building.